+++ src/test/java/kitchenpos/menu/MenuServiceTest.java
@@ -1,26 +1,22 @@
 package kitchenpos.menu;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
-import static org.mockito.Mockito.doThrow;
 
 import java.math.BigDecimal;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-import kitchenpos.menu.application.MenuGroupService;
 import kitchenpos.menu.application.MenuService;
 import kitchenpos.menu.dao.MenuProductRepository;
 import kitchenpos.menu.dao.MenuRepository;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.domain.MenuProducts;
+import kitchenpos.menu.domain.MenuValidator;
 import kitchenpos.menu.dto.MenuProductRequest;
 import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.menu.dto.MenuResponse;
-import kitchenpos.product.application.ProductService;
 import kitchenpos.product.domain.Product;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -40,13 +36,10 @@ class MenuServiceTest {
     MenuRepository menuRepository;
 
     @Mock
-    MenuGroupService menuGroupService;
-
-    @Mock
     MenuProductRepository menuProductRepository;
 
     @Mock
-    ProductService productService;
+    MenuValidator menuValidator;
 
     Menu 후라이드치킨;
     Product 후라이드;
@@ -76,19 +69,6 @@ class MenuServiceTest {
 
         // then
         assertThat(actual.getName()).isEqualTo("후라이드치킨");
-    }
-
-    @Test
-    @DisplayName("메뉴 저장시 메뉴는 존재하는 메뉴그룹 정보를 가지고 있다")
-    void create_nonExistMenuGroupError() {
-        // given
-        doThrow(new IllegalArgumentException("존재하지 않는 메뉴그룹입니다."))
-                .when(menuGroupService).existsById(any());
-
-        // when & then
-        assertThatIllegalArgumentException().isThrownBy(
-                () -> menuService.create(상품)
-        ).withMessageContaining("존재하지 않는 메뉴그룹입니다.");
     }
 
     @Test
